revision 1.4
I believe this reflects what we talked about. The story has been expanded to match your framework. On Shu's request, I've included more on tough times, and the language about the relationship growing like a seed.

It's goddamn late. I'm going to bed. You're a saint for taking this on.

-Matt